fix: cannot root str with different return types

+++ terminal_commander.cpp
@@ -7,7 +7,7 @@
 
 #include "terminal_commander.h"
 
-int16_t strcmp(const char *s1, const char *s2) {
+int strcmp(const char *s1, const char *s2) {
   const unsigned char *p1 = (const unsigned char *)s1;
   const unsigned char *p2 = (const unsigned char *)s2;
